fix several issues detected by llvm

+++ backend/redis/test/backend_redis_resp_parse_test.c
@@ -34,6 +34,7 @@
 #include "../backend/redis/create.h"
 #include "../backend/redis/parse.h"
 #include "../backend/redis/protocol.h"
+#include "../backend/redis/result.h"
 #include "../backend/transports/memcopy.h"
 
 #define DBBE_TEST_BUFFER_LEN ( 1024 )
@@ -58,7 +59,7 @@ int TestNSCreate( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   ":1\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_nscreate( req, &result ), 0 );
@@ -76,7 +77,7 @@ int TestNSCreate( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                       "+OK\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_nscreate( req, &result ), 0 );
@@ -108,7 +109,7 @@ int TestNSAttach( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   ":1\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_nsattach( req, &result ), 0 );
@@ -126,7 +127,7 @@ int TestNSAttach( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   ":5\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_nsattach( req, &result ), 0 );
@@ -161,7 +162,7 @@ int TestNSDetach( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   ":1\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_nsdetach( req, &result ), 0 );
@@ -178,7 +179,7 @@ int TestNSDetach( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   ":5\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_nsdetach( req, &result ), 0 );
@@ -226,7 +227,7 @@ int TestNSDelete( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   ":5\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_nsdelete( &req_io, &result, post_queue, cmr ), 0 );
@@ -247,7 +248,7 @@ int TestNSDelete( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   "*2\r\n$1\r\n0\r\n*5\r\n$3\r\nbla\r\n$2\r\nhi\r\n$5\r\nfasel\r\n$4\r\nfoob\r\n$6\r\ngnartz\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_nsdelete( &req_io, &result, post_queue, cmr ), 0 );
@@ -271,7 +272,7 @@ int TestNSDelete( const char *namespace,
                     dbBE_Redis_sr_buffer_get_size( sr_buf ),
                     ":1\r\n");
     rc += TEST_NOT( len, -1 );
-    rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+    rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
     rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
     rc += TEST( dbBE_Redis_process_nsdelete( &req_io, &result, post_queue, cmr ), 0 );
@@ -294,7 +295,7 @@ int TestNSDelete( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   ":1\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_nsdelete( &req_io, &result, post_queue, cmr ), 0 );
@@ -329,7 +330,7 @@ int TestPut( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   ":1\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_put( req, &result ), 0 );
@@ -360,7 +361,7 @@ int TestRead( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   "$12\r\nReturnString\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_get( req, &result, transport ), 0 );
@@ -406,7 +407,7 @@ int TestDirectory( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   "*3\r\n$6\r\nTestNS\r\n$5\r\ncount\r\n$1\r\n7\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_directory( &req_io, &result, transport, post_queue, cmr ), 0 );
@@ -427,7 +428,7 @@ int TestDirectory( const char *namespace,
                   dbBE_Redis_sr_buffer_get_size( sr_buf ),
                   "*2\r\n$1\r\n0\r\n*5\r\n$11\r\nTestNS::bla\r\n$10\r\nTestNS::hi\r\n$13\r\nTestNS::fasel\r\n$12\r\nTestNS::foob\r\n$14\r\nTestNS::gnartz\r\n");
   rc += TEST_NOT( len, -1 );
-  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), len );
+  rc += TEST( dbBE_Redis_sr_buffer_add_data( sr_buf, len, 0 ), (size_t)len );
 
   rc += TEST( dbBE_Redis_parse_sr_buffer( sr_buf, &result ), 0 );
   rc += TEST( dbBE_Redis_process_directory( &req_io, &result, transport, post_queue, cmr ), 0 );
